upgrade to 0.8.0
Summary: This is part of the first attempt to upgrade our internal Zeppelin to 0.8 (latest public). I tried to rebase all our commits on top of the public version.

The main hotspots in this are:

A couple of js/html files that are kept w/ custom twitter rules (credential.html, home.html, navcontroller.js)
A couple of java classes that hold twitter's custom elfowl-based auth
The idea here is to fix compilation issues and basically sanity-check the whole thing by hand. In a later PR, I'd like to:

Try and extract the whole auth logic into somewhere more digestible/rebase proof.
Explore the possibility of using LDAP instead of elfowl?
Explore the possibility of auto-filling the credential.html from the registered interpreters
I chose this instead of a merge as 1) it keeps the evolution of the internal commits and 2) it starts with a "clean" baseline. If the conflicts here turn out to be too big (I'm specially concerned w/ the js files and how to verify them), we can either 1) squash our internal commits and do a rebase w/ a single pass or 2) merge the public into our repo. Both are equally messy, but 1 is harder to miss, imo.

Differential Revision: https://phabricator.twitter.biz/D59410