[ZEPPELIN-2680] allow opening notebook as a reader
Summary: ### What is this PR for?

We have a few notebooks that have a list of writers & owners, but no restricted readers list.
When you try to open these notebooks as a reader AND they have a chart the app sends a COMMIT_PARAGRAPH for each, which causes a write permission popup to show ("Insufficient privileges to write note").
I tracked down the issue to the result.controller.js (commitParagraphResult) - propagating the "viewOnly" state fixes that behavior.

### What type of PR is it?
Bug Fix

### How should this be tested?

- Create a notebook. Add a paragraph with a graph. Set the "owner" of the notebook to something other than yourself. You won't be able to load the notebook without the security popup anymore (unless logging in as the owner)

### Questions:
* Does the licenses files need update?
No
* Is there breaking changes for older versions?
No
* Does this needs documentation?
No

JIRA Issues: IQ-614

Differential Revision: https://phabricator.twitter.biz/D65473